backend/requirements.txt
fastapi==0.128.2
uvicorn==0.30.6
pdfplumber==0.11.9
pymupdf==1.24.9
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==12.1.1
openai==2.11.0
python-multipart==0.0.22
httpx==0.28.1
pypdf==6.9.2
firebase-admin==7.1.0
google-cloud-storage==3.9.0
google-cloud-kms==3.6.0
google-cloud-tasks==2.21.0
protobuf==5.29.6
PyJWT==2.12.1
reportlab==4.4.4
pyHanko==0.33.0
stripe==13.2.0
pyhanko==0.34.1
